// src/interfaces/Actions/SetViewFilter.tsx
// Generated by https://quicktype.io

export interface SetViewFilter {
    Type:               string;
    ViewFilters:        ViewFilter[];
    Notes:              string;
    ActionId:           number;
    Disabled:           boolean;
    InValid:            boolean;
    ValidationMessages: any[];
    Metadata:           null;
}

export interface ViewFilter {
    ControlId:          string;
    ViewFilterId:       string;
    ValidationMessages: any[];
}